fixes to directories and tests, addition to loader, work in progress for ssh
