+++ honeypot/UnitTest/ssh_plugin_test.py
@@ -3,12 +3,14 @@ __author__ = 'daemoniclegend'
 import os
 import sys
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
-from plugins import ssh_plugin
+path = os.path.dirname(os.path.realpath(__file__)).replace("UnitTest", "plugins")
+sys.path.insert(0, path)
+
+import ssh_plugin
 '''
 path = os.getcwd().replace('UnitTest', '')
 sys.path.insert(0, path)
 sys.path.append('..')
 '''
 
-sshServer = ssh_plugin()
+ssh_plugin.server_plugin()
